refactor(assistant): move generic job-queue infra into persistence/ with handler-registration seam
