+++ assistant/src/memory/worker.ts
@@ -16,6 +16,7 @@ import { getConfig } from "../config/loader.js";
 import { getLogger } from "../util/logger.js";
 import { getMemoryWorkerPidPath } from "../util/platform.js";
 import { startInProcessMemoryJobsWorker } from "./jobs-worker.js";
+import { registerMemoryJobHandlers } from "./register-job-handlers.js";
 
 const log = getLogger("memory-worker-process");
 
@@ -41,6 +42,7 @@ async function main(): Promise<void> {
   writeFileSync(pidPath, String(process.pid), { flag: "w" });
   log.info({ pid: process.pid, pidPath }, "Memory worker process started");
 
+  registerMemoryJobHandlers();
   const worker = startInProcessMemoryJobsWorker();
 
   // Keep-alive: the worker's setTimeout timers are unref'd, so without
